exec image cache reuses pristine ELF frames; console cells widen to 16-bit glyphs
* core/elf.c (elf_load_from_path_info): map a cached pristine image as RO
  Soft_COW instead of re-reading and re-copying the file; skip RELA replay on
  cache hits (snapshots are taken post-relocation); zero unaligned head gap
  and past-EOF tail explicitly since fresh frames are no longer pre-zeroed;
  register the image and demote the first runner after a cold load.
  (kernel_execve_into_mm): skip the redundant open-probe when the path already
  has a snapshot.

* core/mm.c (mm_make_private_range_impl, mm_make_private_range_nozero,
  exec_img_key, exec_img_path_hash, exec_img_path_cached, exec_img_lookup,
  exec_img_map, exec_img_register, exec_img_demote_owner): add Linux
  page-cache style exec image cache -- FNV-64 path/size/window key, bounded
  LRU budget, one frame retain per cached page, snapshot only from the second
  cold load so first-run and boot images keep the old eager behaviour; let
  ELF exec allocate fresh PT_LOAD frames without per-page memset.

* core/process.c (process_reap, process_reap_zombie, process_discard):
  detach the process from all threads before the process_t is freed.

* cpu/thread/thread.c (thread_detach_process): NULL out ->process/linux_tgid
  on every thread still pointing at the dying process so scheduler autoreap
  and thread_wait_parent_alive cannot deref a freed process_t (UAF).
  (thread_wake_expired_timeouts, thread_schedule): merge the zombie-autoreap
  and expired-timeout scans into one pass; the global timer-tick timeout scan
  marks work done so the next schedule skips the duplicate walk.

* cpu/ints/gdt.c (set_user_fs_base): skip the WRMSR when the FS base for this
  CPU is unchanged.

* syscall/dispatch/syscall_dispatch.c (console_kd_reset_for_dead_owner):
  force a full reblit only when a KD/alt-screen transition actually happened,
  saving 2-3 ms on a plain foreground child exit.  (syscall_bind_kstack_
  for_thread): cache last tid/rsp0 per APIC id to avoid redundant stack-top
  writes.  (syscall_maybe_vfork_wait): add a bounded vfork-block duration
  trace.

* drv/video/cirrus/cirrusfb.c, drv/video/vbe/vbefb.c (cell_t): widen the
  text-cell glyph from 8 to 16 bits (glyph16LE + attr) so UTF-8/Unicode
  glyphs survive the backing buffer; snapshot/restore build 4-byte cells.
  drv/video/console.c, drv/video/fonts.c: 16-bit glyph plumbing through the
  blit and literal-put paths; VGA path casts down to a byte glyph.

* fs/virt/devfs.c: back ttys with 4-byte cells (tty_cell_store/tty_cell_glyph);
  UTF-8 fold emits 16-bit glyphs with a '?' fallback; pack down to legacy
  8-bit VGA cells on blit to VRAM.  (devfs_tty_leave_alt_screen): return
  whether an alt-screen transition happened.

* inc/mm.h, inc/thread.h, inc/devfs.h, inc/cirrusfb.h, inc/console.h,
  inc/font.h, inc/vbe.h: match the new signatures and document the exec image
  cache and 4-byte cell layout.

* makefile (run, run-uefi): drop -enable-kvm -cpu host from the plain run
  target and add -vga vmware to the UEFI targets.

+++ drv/video/fonts.c
@@ -117,7 +117,7 @@ static inline void put_pix(uint8_t *line, uint32_t bpp, uint32_t x, uint32_t pix
 }
 
 static void font_blit_glyph_generic(void *fb, uint32_t pitch, uint32_t bpp_bytes,
-				    uint32_t px, uint32_t py, uint8_t ch,
+				    uint32_t px, uint32_t py, uint16_t ch,
 				    uint32_t fg, uint32_t bg) {
 	const struct font *f = font_current();
 	const uint8_t *rows = glyph_rows(ch);
@@ -174,7 +174,7 @@ static void font_fill_cell_32(void *fb, uint32_t pitch, uint32_t px, uint32_t py
 }
 
 void font_blit_glyph_32(void *fb, uint32_t pitch, uint32_t px, uint32_t py,
-			uint8_t ch, uint32_t fg, uint32_t bg) {
+			uint16_t ch, uint32_t fg, uint32_t bg) {
 	const struct font *f = font_current();
 	const uint8_t *rows = glyph_rows(ch);
 	uint32_t w = f->cwidth;
@@ -229,7 +229,7 @@ void font_blit_glyph_32(void *fb, uint32_t pitch, uint32_t px, uint32_t py,
 }
 
 void font_blit_glyph(void *fb, uint32_t pitch, uint32_t bpp_bytes,
-		     uint32_t px, uint32_t py, uint8_t ch,
+		     uint32_t px, uint32_t py, uint16_t ch,
 		     uint32_t fg, uint32_t bg) {
 	if (!fb || bpp_bytes == 0)
 		return;
